requirements.txt
numpy
Pillow
scipy
torch
tqdm
